get SCE for 10k PBMC annotated as in OSCA book chapter 12

get SCE for 10k PBMC annotated as in OSCA book chapter 12

getCh12Sce(clear_cache = FALSE)

Arguments

clear_cache

logical(1) will delete relevant entries in available cache before continuing, defaults to FALSE

Value

SingleCellExperiment instance

Note

This is a SingleCellExperiment instance with data on 7472 cells from a 10x CITE-seq experiment. An altExp component includes antibody-derived tag (ADT) counts on 17 proteins. The data are acquired and processed as described in ch 12 of the OSCA book, circa February 2023. A metadata element (se.averaged) includes the result of averaging protein abundance estimates within ADT-based clusters, as is done to give rise to Figure 12.8 of the OSCA book.

Examples

ch12sce <- getCh12Sce()
ch12sce
#> class: SingleCellExperiment 
#> dim: 33538 7472 
#> metadata(2): Samples se.averaged
#> assays(2): counts logcounts
#> rownames(33538): ENSG00000243485 ENSG00000237613 ... ENSG00000277475
#>   ENSG00000268674
#> rowData names(3): ID Symbol Type
#> colnames(7472): AAACCCAAGATTGTGA-1 AAACCCACATCGGTTA-1 ...
#>   TTTGTTGTCGAGTGAG-1 TTTGTTGTCGTTCAGA-1
#> colData names(3): Sample Barcode sizeFactor
#> reducedDimNames(0):
#> mainExpName: Gene Expression
#> altExpNames(1): Antibody Capture