get an anndata reference to 5k10k protein after totalVI from tutorial

get an anndata reference to 5k10k protein after totalVI from tutorial

getPro5k10kAdata()

Value

python reference to anndata

Note

It may be advantageous to set `options(timeout=3600)` or to allow an even greater time for internet downloads, if working at a relatively slow network connection.

Examples

getPro5k10kAdata()
#> AnnData object with n_obs × n_vars = 10849 × 14
#>     obs: 'n_genes', 'percent_mito', 'n_counts', 'batch', '_scvi_labels', '_scvi_batch', 'leiden_totalVI', 'CD3_fore_prob', 'CD4_fore_prob', 'CD8a_fore_prob', 'CD14_fore_prob', 'CD15_fore_prob', 'CD16_fore_prob', 'CD56_fore_prob', 'CD19_fore_prob', 'CD25_fore_prob', 'CD45RA_fore_prob', 'CD45RO_fore_prob', 'PD-1_fore_prob', 'TIGIT_fore_prob', 'CD127_fore_prob'
#>     var: 'protein_names'
#>     uns: 'dendrogram_leiden_totalVI', 'log1p'
#>     obsm: 'X_totalVI', 'X_umap'